requirement.txt
Flask==2.0.3
python-dotenv==0.19.2
line-bot-sdk==1.20.0
requests==2.26.0
